api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
anthropic==0.7.8
Pillow==10.1.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic[email]==2.5.0
fastapi-cors==0.0.6
